Tests/LibWeb/Text/expected/Worker/Worker-blob.txt

Message received from worker: "WHF! :^)"
